Movement, terrain, and environment update
Player has been added as a simple capsule with the base cs script that allows for movement in all directions using the arrow keys and space key. In addition an environment was added with ambient blue light, volumetric fog and normal fog. The terrain has also been updated to have several hills, plateaus, and a base grass texture.