Environment, character, and terrain update
Environment was added back in with a bluish fog. It is saved in assets so that it may be removed and added back doing development. The player can now be moved with arror keys or WASD. Shift allows for 2x speed but only when traveling in a forward direction. Terrain height has been shortened to deal with player not being able to climb terrain.